% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.imp.pval.R
\name{calc.imp.pval}
\alias{calc.imp.pval}
\title{Calculate Importance p-value}
\usage{
calc.imp.pval(rp, imp.names)
}
\arguments{
\item{rp}{an \code{\link{rfPermute}} object with a \code{null.dist} element.}

\item{imp.names}{vector of column names of an importance metric from the 
\code{importance} and \code{null.dist} elements.}
}
\value{
a vector of p-values for each predictor
}
\description{
Calculate p-values for an importance metric from an rfPermute object. Used internally.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\keyword{internal}

