% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casePredictions.R
\name{casePredictions}
\alias{casePredictions}
\title{Case Predictions}
\usage{
casePredictions(x)
}
\arguments{
\item{x}{a \code{rfPermte} or \code{randomForest} model object.}
}
\value{
A data frame containing columns of original and predicted cases,
  whether they were correctly classified, and vote distributions among cases.
}
\description{
Construct a data frame of case predictions for 
  training data along with vote distributions.
}
\examples{
library(randomForest)
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)

cp <- casePredictions(rf)
cp

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
