% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classConfInt.R
\name{classConfInt}
\alias{classConfInt}
\title{Classification Confidence Intervals}
\usage{
classConfInt(rf, conf.level = 0.95)
}
\arguments{
\item{rf}{a \code{\link[randomForest]{randomForest}} object}

\item{conf.level}{confidence level for the \code{\link{binom.test}} confidence interval}
}
\description{
Calculate confidence intervals for Random Forest classifications
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars, importance = TRUE)
classConfInt(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

