% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanRFdata.R
\name{cleanRFdata}
\alias{cleanRFdata}
\title{Clean Random Forest Input Data}
\usage{
cleanRFdata(x, y, data, max.levels = 30)
}
\arguments{
\item{x}{columns used as predictor variables as character or numeric vector.}

\item{y}{column used as response variable as character or numeric.}

\item{data}{data.frame containing \code{x} and \code{y} columns.}

\item{max.levels}{maximum number of levels in response variable \code{y}.}
}
\value{
a data.frame containing cleaned data.
}
\description{
Removes cases for a Random Forest classification model 
  with missing data and predictors that are constant.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

