% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine rfPermute objects}
\usage{
combine(...)
}
\arguments{
\item{\dots}{two or more objects of class \code{rfPermute}, to be combined 
into one.}
}
\description{
Combines two or more ensembles of \code{rfPermute} objects into 
  one, combining \code{randomForest} results, null distributions, 
  and re-calculating p-values.
}
\note{
This function has the same name as \code{randomForest::combine} 
  so it might be masked by that function if \code{randomForest} is loaded 
  after \code{rfPermute}. If all objects in \code{...} are pure 
  \code{randomForest} models, this function will just run 
  \code{randomForest::combine} and return the results.
}
\examples{
data(iris)
rp1 <- rfPermute(
  Species ~ ., iris, ntree = 50, norm.votes = FALSE, nrep = 100, num.cores = 1
)
rp2 <- rfPermute(
  Species ~ ., iris, ntree = 50, norm.votes = FALSE, nrep = 100, num.cores = 1
)
rp3 <- rfPermute(
  Species ~ ., iris, ntree = 50, norm.votes = FALSE, nrep = 100, num.cores = 1
)
rp.all <- combine(rp1, rp2, rp3)
rp.all

plotNull(rp.all) 

}
\seealso{
\code{\link[randomForest]{combine}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
