% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\alias{plotConfMat}
\title{Confusion Matrix}
\usage{
confusionMatrix(x, conf.level = 0.95, threshold = NULL, digits = 1)

plotConfMat(x, title = NULL, plot = TRUE)
}
\arguments{
\item{x}{a \code{rfPermute} or \code{randomForest} model object.}

\item{conf.level}{confidence level for the \code{\link{binom.test}} confidence interval}

\item{threshold}{threshold to test observed classification probability against.}

\item{digits}{integer indicating the number of decimal places to round values 
to.}

\item{title}{a title for the plot.}

\item{plot}{display the plot?}
}
\description{
Generate a confusion matrix for Random Forest analyses with 
  error rates translated into percent correctly classified, and columns for 
  confidence intervals and expected classification rates (priors) added.
}
\examples{
library(randomForest)
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars, importance = TRUE)
confusionMatrix(rf)

}
\seealso{
\code{\link{plotConfMat}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
