% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exptd.err.rate.R
\name{exptd.err.rate}
\alias{exptd.err.rate}
\title{Expected Error Rate}
\usage{
exptd.err.rate(rf)
}
\arguments{
\item{rf}{an object inheriting from \code{link{randomForest}}.}
}
\value{
a vector of expected error rates for each class.
}
\description{
Calculate expected OOB error rates for randomForest 
  classification model based on random assignment and class sizes (prior).
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

