% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exptdErrRate.R
\name{exptdErrRate}
\alias{exptdErrRate}
\title{Expected Error Rate}
\usage{
exptdErrRate(rf)
}
\arguments{
\item{rf}{an object inheriting from \code{link{randomForest}}.}
}
\value{
a matrix of expected error rates (\code{prior}) and 
  p-values from a binomial test (\code{class_p.value}) for each class.
}
\description{
Calculate expected OOB error rates (priors) for randomForest 
  classification model based on random assignment and class sizes.
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
exptdErrRate(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
