% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impHeatmap.R
\name{impHeatmap}
\alias{impHeatmap}
\title{Importance Heatmap}
\usage{
impHeatmap(rf, ranks = TRUE)
}
\arguments{
\item{rf}{an object inheriting from \code{link{randomForest}}.}

\item{ranks}{plot ranks instead of actual importance scores?}
}
\description{
Plot heatmap of importance scores or ranks
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

