% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\alias{importance.rfPermute}
\alias{plotImportance}
\title{Extract rfPermute Importance Scores and p-values.}
\usage{
\method{importance}{rfPermute}(x, scale = TRUE, sort.by = NULL, decreasing = TRUE, ...)

plotImportance(
  x,
  scale = TRUE,
  alpha = 0.05,
  sig.only = FALSE,
  type = NULL,
  n = NULL,
  main = NULL
)
}
\arguments{
\item{x}{An object produced by a call to \code{rfPermute}.}

\item{scale}{For permutation based measures, should the measures be divided 
their "standard errors"?}

\item{sort.by}{character vector giving the importance metric(s) or p-values 
to sort by. If \code{NULL}, defaults to \code{"MeanDecreaseAccuracy"} for 
classification models and \code{"\%IncMSE"} for regression models.}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{...}{arguments to be passed to and from other methods.}

\item{alpha}{Critical alpha to identify "significant" predictors.}

\item{sig.only}{Plot only the significant (<= \code{alpha}) predictors?}

\item{type}{character vector listing which importance measures to plot.
Can be class names or names of overall importance measures 
(e.g., "MeanDecreaseAccuracy").}

\item{n}{Plot \code{n} most important predictors.}

\item{main}{Main title for plot.}
}
\description{
Extract a matrix of the observed importance scores
  and p-values from the object produced by a call to \code{rfPermute}
}
\details{
p-values can be given to the \code{sort.by} argument by adding 
  '.pval' to the column name of the desired column from the \code{importance} 
  element of the \code{rfPermute} object.
}
\examples{
data(mtcars)

# A classification model classifying cars to manual or automatic transmission 
am.rp <- rfPermute(factor(am) ~ ., mtcars, ntree = 100, nrep = 50)
  
imp.scaled <- importance(am.rp, scale = TRUE)
imp.scaled

# plot scaled importance scores
plotImportance(am.rp, scale = TRUE)

# plot unscaled and only significant scores
plotImportance(am.rp, scale = FALSE, sig.only = TRUE)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
