% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBag.R
\name{plotBag}
\alias{plotBag}
\title{Plot inbag or OOB distribution}
\usage{
plotBag(x, type = c("inbag", "oob"), sampsize = NULL, bins = NULL, plot = TRUE)
}
\arguments{
\item{x}{a \code{rfPermute} or \code{randomForest} model object..}

\item{type}{plot the frequency samples were inbag (\code{"inbag"}) or
out-of-bag (\code{"oob"}).}

\item{sampsize}{optional vector of sample sizes used in \code{rf} model.}

\item{bins}{number of bins in histogram.}

\item{plot}{display the plot?}
}
\value{
the \code{ggplot2} object is invisibly returned. If \code{sampsize} 
  is not \code{NULL} and \code{rf} contains a classification model, the plot 
  will display red vertical lines at the expected values for the classes 
  based on their frequency and sample sizes.
}
\description{
Plot distributions of the fraction of trees that samples were
  inbag or out-of-bag (OOB) in the Random Forest model.
}
\examples{
library(randomForest)
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
plotBag(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
