% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConfMat.R
\name{plotConfMat}
\alias{plotConfMat}
\title{Plot Confusion Matrix}
\usage{
plotConfMat(rf, title = NULL, plot = TRUE)
}
\arguments{
\item{rf}{an object inheriting from \code{\link{randomForest}}.}

\item{title}{a title for the plot.}

\item{plot}{display the plot?}
}
\value{
the \code{ggplot2} object is invisibly returned.
}
\description{
Plot confusion matrix heatmap.
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
plotConfMat(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
