% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpHeatmap.R
\name{plotImpHeatmap}
\alias{plotImpHeatmap}
\title{Importance Heatmap}
\usage{
plotImpHeatmap(
  x,
  n = NULL,
  ranks = TRUE,
  plot = TRUE,
  xlab = NULL,
  ylab = NULL,
  scale = TRUE,
  alpha = 0.05,
  size = 3
)
}
\arguments{
\item{x}{a \code{rfPermute} or \code{randomForest} model object.}

\item{n}{plot \code{n} most important predictors.}

\item{ranks}{plot ranks instead of actual importance scores?}

\item{plot}{print the plot?}

\item{xlab, ylab}{labels for the x and y axes.}

\item{scale}{for permutation based measures, should the measures be divided 
their "standard errors"?}

\item{alpha}{a number specifying the critical alpha for identifying 
predictors with importance scores significantly different from random. 
This parameter is only relevant if \code{rf} is a \code{\link{rfPermute}}
object with p-values. Importance measures with p-values less than or 
equal to \code{alpha} will be denoted in the heatmap by a white diamond. 
If set to \code{NULL}, significance is not denoted.}

\item{size}{a value specifying the size of the significance diamond if 
the p-value <= \code{alpha}.}
}
\value{
the \code{ggplot} object is invisibly returned.
}
\description{
Plot heatmap of importance scores or ranks from a 
  classification model
}
\details{
\code{rf} must be a classification model run with 
  \code{importance = TRUE}.
}
\examples{
data(mtcars)

# A randomForest model
rf <- randomForest::randomForest(factor(am) ~ ., mtcars, importance = TRUE)
importance(rf)
plotImpHeatmap(rf, xlab = "Transmission", ylab = "Predictor")

# An rfPermute model with significant predictors identified
rp <- rfPermute(factor(am) ~ ., mtcars, nrep = 100, num.cores = 1)
importance(rp)
plotImpHeatmap(rp, xlab = "Transmission", ylab = "Predictor")

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
