% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOOBtimes.R
\name{plotOOBtimes}
\alias{plotOOBtimes}
\title{Plot Times OOB}
\usage{
plotOOBtimes(rf, bins = NULL, plot = TRUE)
}
\arguments{
\item{rf}{an object inheriting from \code{\link{randomForest}}.}

\item{bins}{number of bins in histogram. Defaults to number of samples / 5.}

\item{plot}{display the plot?}
}
\value{
the \code{ggplot2} object is invisibly returned.
}
\description{
Plot histogram of times samples were OOB.
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
plotOOBtimes(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
