% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRFtrace.R
\name{plotRFtrace}
\alias{plotRFtrace}
\title{OOB Trace}
\usage{
plotRFtrace(rf, plot = TRUE)
}
\arguments{
\item{rf}{an object inheriting from \code{\link{randomForest}}.}

\item{plot}{display the plot?}
}
\value{
the \code{ggplot2} object is invisibly returned.
}
\description{
Plot trace of cumulative OOB error rate by number of trees
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
plotRFtrace(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
