% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfPermute_package.R
\docType{package}
\name{rfPermute_package}
\alias{rfPermute_package}
\alias{rfPermuteTutorial}
\alias{rfPermute-package}
\title{\code{rfPermute} package}
\usage{
rfPermuteTutorial()
}
\description{
Random Forest Predictor Importance Significance and Model Diagnostics.
}
\keyword{package}
