% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfROC.R
\name{rfROC}
\alias{rfROC}
\title{Receiver Operator Curve}
\usage{
rfROC(x, plot = TRUE)
}
\arguments{
\item{x}{a \code{rfPermute} or \code{randomForest} model object.}

\item{plot}{logical? Plot the curve?}
}
\value{
\describe{ a list returned invisbly containing:
 \item{\code{roc}}{a list containing ROC results from \code{\link[pROC]{roc}} for each class}
 \item{\code{smry}}{a data fame summarizing the threshold and AUC for each class}
 \item{\code{plot}}{the \code{ggplot} plot of the ROC curves}
}
}
\description{
Create a Receiver Operator Curve (ROC) for each class in a 
  Random Forest model and (optionally) plot them.
}
\examples{
library(randomForest)
data(mtcars)

# random sampling with replacement
rf <- randomForest(factor(am) ~ ., mtcars)

roc <- rfROC(rf)

roc$smry

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
