% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.imp.pval}
\alias{calc.imp.pval}
\title{Calculate Importance p-value}
\usage{
calc.imp.pval(rp, imp.type)
}
\arguments{
\item{rp}{an \code{\link{rfPermute}} object with a \code{null.dist} element.}

\item{imp.type}{column name of an importance metric from the \code{importance} and \code{null.dist} elements.}
}
\value{
a vector of p-values for each predictor
}
\description{
Calculate p-values for an importance metric from an rfPermute object. Used internally.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

