% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casePredictions.R
\name{casePredictions}
\alias{casePredictions}
\title{Case Predictions}
\usage{
casePredictions(rf)
}
\arguments{
\item{rf}{an object inheriting from \code{\link{randomForest}}.}
}
\value{
A data frame containing columns of original and predicted cases,
  whether they were correctly classified, and vote distributions.
}
\description{
Get data frame of case predictions for training data along with 
  vote distributions.
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
casePredictions(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
