% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean.rf.data}
\alias{clean.rf.data}
\title{Clean Random Forest Input Data}
\usage{
clean.rf.data(x, y, data, max.levels = 30)
}
\arguments{
\item{x}{columns used as predictor variables as character or numeric vector.}

\item{y}{column used as response variable as character or numeric.}

\item{data}{data.frame containing 'x' and 'y' columns.}

\item{max.levels}{maximum number of levels in response variable 'y'.}
}
\value{
a data.frame containing cleaned data.
}
\description{
Removes cases with missing data and predictors that are constant.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

