% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exptd.err.rate.R
\name{exptd.err.rate}
\alias{exptd.err.rate}
\title{Expected Error Rate}
\usage{
exptd.err.rate(rf)
}
\arguments{
\item{rf}{a \code{link{randomForest}} or \code{\link{rfPermute}} object.}
}
\value{
a vector of expected error rates.
}
\description{
Calculate expected OOB error rates based on random assignment
  and class sizes (prior).
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

