% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impHeatmap.R
\name{impHeatmap}
\alias{impHeatmap}
\title{Importance Heatmap}
\usage{
impHeatmap(rf, ranks = TRUE, plot = TRUE, xlab = NULL, ylab = NULL)
}
\arguments{
\item{rf}{an object inheriting from \code{link{randomForest}}.}

\item{ranks}{plot ranks instead of actual importance scores?}

\item{plot}{print the plot?}

\item{xlab, ylab}{labels for the x and y axes.}
}
\value{
the \code{ggplot} object is invisibly returned.
}
\description{
Plot heatmap of importance scores or ranks from a 
  classification model
}
\details{
\code{rf} must be a classification model run with 
  \code{importance = TRUE}.
}
\examples{
data(mtcars)
rf <- randomForest(factor(am) ~ ., mtcars, importance = TRUE)
impHeatmap(rf, xlab = "Transmission", ylab = "Predictor")

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

