% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rp.importance.R
\name{plot.rp.importance}
\alias{plot.rp.importance}
\title{Plot Random Forest Importance Distributions}
\usage{
\method{plot}{rp.importance}(x, alpha = 0.05, sig.only = FALSE,
  type = NULL, n = NULL, main = NULL, ...)
}
\arguments{
\item{x}{An object produced by a call to \code{\link{rp.importance}}.}

\item{alpha}{Critical alpha to identify "significant" predictors.}

\item{sig.only}{Plot only the significant (<= \code{alpha}) predictors?}

\item{type}{character vector listing which importance measures to plot.
Can be class names or names of overall importance measures 
(e.g., "MeanDecreaseAccuracy") in the \code{\link{rp.importance}} object.}

\item{n}{Plot \code{n} most important predictors.}

\item{main}{Main title for plot.}

\item{...}{Optional arguments which will be ignored.}
}
\description{
Plot the Random Forest importance distributions,
with significant p-values as estimated in rfPermute.
}
\details{
The function will generate a panel of plots, one for each 
  importance type.
}
\examples{
# A regression model using the ozone example
data(airquality)
ozone.rfP <- rfPermute(
  Ozone ~ ., data = airquality, ntree = 100, 
  na.action = na.omit, nrep = 50, num.cores = 1
)
  
# Plot the unscaled importance distributions and highlight significant predictors
plot(rp.importance(ozone.rfP, scale = FALSE))
  
# ... and the scaled measures
plot(rp.importance(ozone.rfP, scale = TRUE))

}
\seealso{
\code{\link{rfPermute}}, \code{\link{rp.importance}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
