% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInbag.R
\name{plotInbag}
\alias{plotInbag}
\title{Plot Inbag distribution}
\usage{
plotInbag(x, bins = 10, replace = TRUE, sampsize = NULL, plot = TRUE)
}
\arguments{
\item{x}{a \code{rfPermute} or \code{randomForest} model object..}

\item{bins}{number of bins in histogram.}

\item{replace}{was sampling done with or without replacement?}

\item{sampsize}{sizes of samples drawn. Either a}

\item{plot}{display the plot?}
}
\value{
the \code{ggplot2} object is invisibly returned. If \code{sampsize} 
  is not \code{NULL} and \code{rf} contains a classification model, the plot 
  will display red vertical lines at the expected values for the classes 
  based on their frequency and sample sizes.
}
\description{
Plot distributions of the fraction of trees that samples were
  inbag in the Random Forest model.
}
\examples{
library(randomForest)
data(mtcars)

sampsize = c(5, 5)

rf <- randomForest(factor(am) ~ ., data = mtcars, ntree = 10)
plotInbag(rf)

rf <- randomForest(factor(am) ~ ., data = mtcars, ntree = 1000)
plotInbag(rf)

rf <- randomForest(factor(am) ~ ., data = mtcars, ntree = 10000)
plotInbag(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
