% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVotes.R
\name{plotVotes}
\alias{plotVotes}
\title{Vote Distribution}
\usage{
plotVotes(rf, plot = TRUE)
}
\arguments{
\item{rf}{an object inheriting from \code{\link{randomForest}}.}

\item{plot}{print the plot?}
}
\value{
the \code{ggplot} object is invisibly returned.
}
\description{
Plot distribution of votes for each sample in each class.
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
plotVotes(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

