% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVotes.R
\name{plotVotes}
\alias{plotVotes}
\title{Vote Distribution}
\usage{
plotVotes(rf, type = NULL, plot = TRUE)
}
\arguments{
\item{rf}{an object inheriting from \code{\link{randomForest}}.}

\item{type}{either \code{area} for stacked continuous area plot or 
\code{bar} for discrete stacked bar chart. The latter is prefered for small 
numbers of cases. If not specified, a bar chart will be used if all 
classes have <= 30 cases.}

\item{plot}{display the plot?}
}
\value{
the \code{ggplot2} object is invisibly returned.
}
\description{
Plot distribution of votes for each sample in each class.
}
\examples{
data(mtcars)

rf <- randomForest(factor(am) ~ ., mtcars)
plotVotes(rf)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
