% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfPermute-package.R
\docType{package}
\name{rfPermute-package}
\alias{rfPermute-package}
\title{\code{rfPermute} package}
\description{
Random Forest Predictor Importance Significance and Model Diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWFSC/rfPermute}
  \item Report bugs at \url{https://github.com/SWFSC/rfPermute/issues}
}

}
\author{
\strong{Maintainer}: Eric Archer \email{eric.ivan.archer@gmail.com}

}
\keyword{internal}
