% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfPermute_package.R
\docType{package}
\name{rfPermute_package}
\alias{rfPermute_package}
\alias{rfPermuteTutorial}
\title{rfPermute model object}
\usage{
rfPermuteTutorial()
}
\description{
Container for rfPermute results
}
\keyword{package}
